streaming/requirements.txt
asyncio
websockets
kafka-python
pyspark